respect opts.useGlobalNPM in sync script
