+++ index.js
@@ -2,7 +2,6 @@ var ValidationError = require('error/validation');
 var find = require('array-find');
 var path = require('path');
 var fs = require('fs');
-var child = require("child_process");
 var sortedObject = require('sorted-object');
 var readJSON = require('read-json');
 
@@ -13,6 +12,7 @@ var walkDeps = require('./walk-shrinkwrap.js');
 var trimNested = require('./trim-nested.js');
 var sync = require('./sync/');
 var ERRORS = require('./errors.js');
+var loadNPM = require('./load-npm.js');
 
 /*  npm-shrinkwrap algorithm
 
@@ -291,43 +291,6 @@ function npmShrinkwrap(opts, callback) {
 }
  
 module.exports = npmShrinkwrap;
-
-/*  you cannot call `npm.load()` twice with different prefixes.
-    
-    The only fix is to clear the entire node require cache and
-      get a fresh duplicate copy of the entire npm library
-*/
-function loadNPM(useGlobalNPM, opts, cb) {
-    Object.keys(require.cache).forEach(function (key) {
-        delete require.cache[key];
-    });
-
-    if (!useGlobalNPM) {
-      require('npm').load(opts, cb);
-
-      return;
-    }
-    
-    child.exec("npm prefix -g", function (error, stdout, stderr) {
-      var _err = stderr || error;
-      if (_err) {
-          // fake error like npm error
-          throw ERRORS.NPMError({
-              pkginfo: _err,
-              problemsText: _err.stack || _err.toString()
-          });
-      }
-
-      // removing trailing new line
-      var prefix = stdout.replace(/\n+$/, "");
-      var npmPath = path.join(prefix, "lib/node_modules/npm");
-      if (fs.existsSync(npmPath)) {
-          require(npmPath).load(opts, cb);
-      } else {
-          throw new Error("global NPM not found in " + npmPath);
-      }
-    });
-}
 
 function NPMError(pkginfo) {
     var problemsText = pkginfo.problems.join('\n');
